requirements.txt
Flask==3.0.0
torch
torchvision
transformers
Pillow
werkzeug